Changed file structure a bit and removed the Transition Group
