+++ src/components/CharList/index.js
