+++ src/components/Character/Character-view.js
