+++ src/components/LoadingScreen/LoadingScreen-view.js
