+++ src/components/SearchBar/SearchBar-view.js
