+++ src/index.css
@@ -1,12 +1,12 @@
 /* CUSTOM FONTS */
 @font-face {
   font-family: 'Jedi Solid';
-  src: url('./fonts/Starjedi.ttf') format('truetype');
+  src: url('./assets/fonts/Starjedi.ttf') format('truetype');
 }
 @font-face {
   font-family: 'Anakin';
   font-weight: 400;
-  src: url('./fonts/anakinmono.ttf') format('truetype');
+  src: url('./assets/fonts/anakinmono.ttf') format('truetype');
 }
